Document unattended operation and correct stale instructions
Adds an operations section covering the watchdog, upload spool, OTA, retention
job, alert scheduling, and the LAN trust boundary.

Corrects three places where the README described behaviour the code no longer
has: the 30-second sampling default, the claim that defining BATTERY_SENSE_PIN
does nothing, the hardcoded serial port paths, and calibration as a
reflash-required edit to main.cpp.

+++ src/main.cpp
@@ -636,8 +636,10 @@ void handleApiData() {
   json += ",\"calibration_dry\":" + String(calibrationDryRaw);
   json += ",\"calibration_wet\":" + String(calibrationWetRaw);
   json += ",\"calibrated\":" + String(calibrationFromUser ? "true" : "false");
-  json += ",\"cloud_pending\":" + String(spoolPendingCount());
-  json += ",\"cloud_dropped\":" + String(spoolDroppedCount);
+  json += ",\"cloud_pending\":" +
+          String(static_cast<unsigned long>(spoolPendingCount()));
+  json += ",\"cloud_dropped\":" +
+          String(static_cast<unsigned long>(spoolDroppedCount));
   json += ",\"cloud_buffer\":\"" + String(spoolReady ? "flash" : "ram") + "\"";
   json += ",\"cloud_online\":" +
           String(cloudUploadOnline ? "true" : "false");
